Resolved NicEastVillage's comments, and improved mapper
Added name to more production parts.
Changed arrayLookupExpr to use a expression as array index.
Changed AST nodes to used stronger types for the ctx variables.
Improved traversal of literals.
Changed increment and decrement statements to takes arbitrary
expressions instead of variable indetifier.
Changed become to take a expression instead of a fixed state.
Removed unneccesary filtering in visitCodeBlock.
Removed the ability to use a semicolon as an empty statement.