Added type checker
Support for implicit conversions are missing.
Added float and integer tokens to grammar.
Changed the TypedNode interface to allow for more classes to contain a
type, and to have a generic way of accessing that type.
Move type of literals to their class definition.
Deprecated SymbolTable in favor of the new CreatingSymbolTableSession.
Removed typeFromString in favor of typeFromCtx.
Added ScopedASTVisitor to traverse the AST while automaticly opening and
closing scopes.
Updated test to include ParseTreeValueWalker, SymbolWalker, and
TypeChecker.